[google_maps_flutter] Remove legacy renderer tests
Removes the copy of the integration tests that attempted to run against
the legacy renderer, as the legacy renderer is no longer available. The
tests are now just duplicating integration tests (the slowest and
flakiest tests, generally), and the test that asserts that it's testing
the legacy renderer is now failing, closing the tree.

Also removes the README discussion of the renderer option, since it's no
longer useful to clients.

Will fix the tree closure.

Part of https://github.com/flutter/flutter/issues/171338
