+++ packages/google_maps_flutter/google_maps_flutter_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.16.2
+
+* Removes documentation related to the map renderer selection API, as the
+  legacy renderer is no longer available, so requesting it is a no-op.
+
 ## 2.16.1
 
 * Removes obsolete code related to supporting SDK <21.
